--- src/features/todo/models.d.ts
@@ -1,13 +1,0 @@
-/** TodoMVC model definitions **/
-
-declare interface TodoItemData {
-  id?: TodoItemId;
-  text?: string;
-  completed?: boolean;
-}
-
-declare type TodoItemId = number;
-
-declare type TodoFilterType = 'SHOW_ALL' | 'SHOW_ACTIVE' | 'SHOW_COMPLETED';
-
-declare type TodoStoreState = TodoItemData[];
