+++ src/features/todo/todos/store.ts
@@ -1,0 +1,7 @@
+export interface TodoData {
+  id?: number;
+  text?: string;
+  completed?: boolean;
+}
+
+export type TodoStore = TodoData[];
