--- src/middleware/logger.ts
@@ -1,6 +1,0 @@
-export default function loggerMiddleware(store) {
-  return next => action => {
-    console.log(action);
-    return next(action);
-  };
-}
